init moving helpers from test to helpers, - testing them separately in helpers_test - initializeLCD should be moved there too - we will launch main for our pre-configured server
